+++ greenfield/eldenring/data.py
@@ -5531,4 +5531,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:3cf291e3385c7a6698e2daeffa7f1509b40bf01f2ece8988865750c286825241', 'module': 'data.py', 'body_sha256': 'sha256:36c6ff4721120de81088192cb0367451deb6b09e2fd408d3dab9f2161d0a3a7a'}
+_GEN_STAMP = {'inputs_hash': 'sha256:8b3e6d2df417e11b5eebf3de01d1a777f8e2cc1141ee1b58d7c032a33048f9d7', 'module': 'data.py', 'body_sha256': 'sha256:36c6ff4721120de81088192cb0367451deb6b09e2fd408d3dab9f2161d0a3a7a'}